step3 code  ver2
